Moduel.js Update
Updated the FFMPEG_Args object. Changed input source to an AviSync .mux
file that is to be located in the Program Files /FFMPEG/bin directory.
This .mux file references am_audio.grf and am_video.grf. These grf files
should also be located in the Program Files/FFMPEG/bin directory. The
grf files are specifically for use with the AmaRecTV application.